set `JOBS` default to the core count
I don't have 8 cores on my laptop, so whenever I want to build the
backend I have to manually change `api/deps.sh` to reduce the `JOBS`
count. This patch sets the default `JOBS` value to the number of cores
on the system, and allows one to configure this value via the `JOBS`
environment variable.